Financial functions moved to a new repository
Hi Marius, 

The financial functions in NumPy are deprecated and provided to the community as a separate package with the name numpy_financial. 
See https://numpy.org/neps/nep-0032-remove-financial-functions.html

I have been working with tm1 for several years but I'm just starting with Python and TM1py. Glad to be able to contribute to the community. Good work Marius !

Cheers,
Albin,